refactor: app sidebar structure and navigation menu
